feat: サイドバー動作設定と画面制御タブの分離
- 採点・解答用紙作成画面でのサイドバー動作を設定可能に（縮小/展開/変更しない）
- 設定はlocalStorageに保存
- 画面制御（プロジェクターモード・画面消灯・サイドバー動作）を独立タブに分離
- 表示設定タブは色設定のみに整理
- 設定ページを4タブ構成に変更（キーボード・画面制御・表示設定・ユーザー管理）
